Fix localStorage support
